Modularize chordname parser
- set root font size using markup (not properties)
- combine root and rest into concat markup (not separate stencils)
- return markup and unprocessed part of input
- remove markup?->stencil?

Separate markup interpretation for root and harmonic specialization is not
necessary. The font size of the root can be increased using markup. Both can
then be combined into one concat markup which is returned in a pair along with
the unprocessed remainder of the input, allowing for integration into more
complex parsing procedures.