Rename chord attach function
Don't drop existing articulations like ties. This implementation is rather
cumbersome, because it needs to access the articulations property twice: once
for retrieving the current value to combine with the markup, and again to change
it.

The value of articulations is apparently not guaranteed to be a list, so using
set-cdr! . last-pair to change the value using only one expression fails. Still,
even in such cases, the resulting pair with markup doesn't cause problems.